pyproject.toml
[project]
name = "task-manager-api"
version = "0.1.0"
description = "A Task Manager API built with FastAPI"
readme = "README.md"
authors = [
    { name = "sagivt1", email = "sagiv.talker@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.118.3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
]

[project.scripts]
task-manager-api = "task_manager_api:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
